Update nokogiri to at least 1.6.7
Nokogiri in ruby 2.2 does not support Windows, and vagrant recently just
merged a pull request to update the dependency to 1.6.7
(See https://github.com/mitchellh/vagrant/pull/6848)
This update enables people to use this plugin on windows with the most recent
versions of vagrant. This should wait to be merged until the patch for vagrant
is released.